05/01/23 - Added an ammo counter to the bottom of the screen, which works by using a counter in the projectile script, and added a working reload bool in the player shoot script

+++ TopDownShooter (Redone)/Assets/Scripts/SO Scripts/BulletDataSO.cs
@@ -16,5 +16,5 @@ public class BulletDataSO : ScriptableObject
     [Header("Bullet Base Stat Multipliers")]
     public int BulletDamageMult;
     public float BulletVelocityMult;
-    public float BulletAmountMult;
+    public int BulletAmountMult;
 }
